### 2023-06-27 #5597 @arfon
Note to editors, a large fraction of the code here is in the GUI-layer. See a more detailed breakdown of the Python code here:

```
arfon@Arfons-MBP ~/sonoUno-desktop/sonoUno (master)$ cloc data_export/
       2 text files.
       2 unique files.
       0 files ignored.

Language                     files          blank        comment           code
Python                           2             20             61            146
SUM:                             2             20             61            146
arfon@Arfons-MBP ~/sonoUno-desktop/sonoUno (master)$ cloc data_import/
       2 text files.
       2 unique files.
       0 files ignored.

Language                     files          blank        comment           code
Python                           2             31            105            140
SUM:                             2             31            105            140
arfon@Arfons-MBP ~/sonoUno-desktop/sonoUno (master)$ cloc data_transform/
       4 text files.
       4 unique files.
       0 files ignored.

Language                     files          blank        comment           code
Python                           4             55            160            144
SUM:                             4             55            160            144
arfon@Arfons-MBP ~/sonoUno-desktop/sonoUno (master)$ cloc sound_module/
       2 text files.
       2 unique files.
       2 files ignored.

Language                     files          blank        comment           code
Python                           2             57             42            224
SUM:                             2             57             42            224
arfon@Arfons-MBP ~/sonoUno-desktop/sonoUno (master)$ cloc gui/
       3 text files.
       3 unique files.
       0 files ignored.

Language                     files          blank        comment           code
Python                           3            714           1480           5566
SUM:                             3            714           1480           5566
```
