+++ paper/paper.bib
@@ -2,7 +2,7 @@
   author = {Díaz-Merced, W. L.}, 
   title = {Sound for the Exploration of Space Physics Data}, 
   booktitle = {University of Glasgow}, 
-  year = {2013} 
+  year = {2013}
 }
 
 @inproceedings{bardellietal2021,
@@ -11,7 +11,8 @@
   booktitle={In International Conference on Human-Computer Interaction},
   pages={171-188},
   year={2021},
-  organization={Springer, Cham}
+  organization={Springer, Cham},
+  doi = {10.1007/978-3-030-77411-0_12}
 }
 
 @inproceedings{lightsound2020,
@@ -20,7 +21,8 @@
   booktitle={In American Astronomical Society Meeting Abstracts\# 235},
   pages={203-204},
   year={2020},
-  organization={}
+  organization={},
+  adsurl = {https://ui.adsabs.harvard.edu/abs/2020AAS...23520304B}
 }
 
 @article{henriquesetal2014,
@@ -30,7 +32,8 @@
   volume = {7},
   number = {1},
   pages = {59-77},
-  year = {2014}
+  year = {2014},
+  doi = {10.4018/jitr.2014010105}
 }
 
 @article{laurentizetal2021,
@@ -41,7 +44,8 @@
   number = {1},
   pages = {13-26},
   year = {2021},
-  note = {Recuperado de: \url{https://www.igi-global.com/article/r-scuti/277414} (accedido:23 de febrero 2023)}
+  note = {Recuperado de: \url{https://www.igi-global.com/article/r-scuti/277414} (accedido:23 de febrero 2023)},
+  doi = {10.4018/IJCICG.2021010102}
 }
 
 @article{reynagaetal2020,
@@ -51,7 +55,8 @@
   volume = {},
   number = {},
   pages = {26-37},
-  year = {2020}
+  year = {2020},
+  doi = {10.4018/978-1-5225-8539-8.ch002}
 }
 
 @inproceedings{andreopoulouYgoudarzi2021,
@@ -60,7 +65,8 @@
   booktitle={In the 26th International Conference on Auditory Display (ICAD), Virtual Conference},
   pages={},
   year={2021},
-  organization={}
+  organization={},
+  doi = {10.21785/icad2021.031}
 }
 
 @inproceedings{fergusonYbrewster2017,
@@ -69,7 +75,8 @@
   booktitle={In Proceedings of the 19th ACM International Conference on Multimodal Interaction},
   pages={120-127},
   year={2017},
-  organization={}
+  organization={},
+  doi = {10.1145/3136755.3136783}
 }
 
 @inproceedings{ise2a2017,
@@ -79,7 +86,8 @@
 	volume={200},
 	pages={01013},
 	year={2019},
-	organization={EDP Sciences}
+	organization={EDP Sciences},
+	doi = {10.1051/epjconf/201920001013}
 }
 
 @article{ijskd2022,
@@ -89,7 +97,8 @@
   volume = {14},
   number = {1},
   pages = {1-27},
-  year = {2022}
+  year = {2022},
+  doi = {10.4018/IJSKD.299048}
 }
 
 @inproceedings{casadoetal2019,
@@ -108,7 +117,8 @@
   volume = {9},
   number = {4},
   pages = {42-51},
-  year = {2022}
+  year = {2022},
+  doi = {10.11648/j.ajaa.20210904.11}
 }
 
 @article{zanella2022,
@@ -119,7 +129,8 @@
   number = {},
   pages = {1-8},
   year = {2022},
-  publisher = {}
+  publisher = {},
+  doi = {10.1038/s41550-022-01721-z}
 }
 
 @article{starsound2022,
[…]
