### 2023-06-27 #5597 @johicasado
> @johicasado – many thanks for your submission to JOSS. A few housekeeping issues here:
>
> * Please move the LICENSE file to the top of the repository.
> * Please review the DOI suggestions from @editorialbot and add them to your BibTeX file if they look to be correct.
>
> Also (a more major comment), please substantially shorten your paper. Our guideline is that JOSS papers should be _up to_ 1000 words long, and yours is well over that. Additionally, there isn't a clear statement of need section that articulates the research need of this package (see our paper preparation guidelines for help here: https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)
>
> Finally, I'm going to trigger a scope review by the JOSS editorial team as it's not immediately clear to me whether this software is in scope for JOSS.

Hello, thank you very much! I already update:
- The LICENSE to the top of the repository
- The DOIs, only two or three references don't present one because it is a thesis and webpages.

About the paper, I will work on it in a short time to accomplish the 1000 words, sorry about that.
