+++ paper/paper.bib
@@ -5,6 +5,16 @@
   year = {2013}
 }
 
+@inproceedings{wandaetal2019, 
+  author = {Díaz-Merced, W. and Walker-Holmes, M. and Hettlage, C. and Green, P. and Casado, J. and García, B.}, 
+  title = {A User Centred Inclusive Web Framework for Astronomy}, 
+  booktitle={Proceedings of the International Astronomical Union},
+  pages={421-422},
+  year={2019},
+  organization={Cambridge University Press},
+  doi = {10.1017/S1743921321000272}
+}
+
 @inproceedings{bardellietal2021,
   author={Bardelli, S. and Ferretti, C. and Ludovico, L. A. and Presti, G. and Rinaldi, M.},
   title={A Sonification of the zCOSMOS Galaxy Dataset},
