+++ paper/paper.md
@@ -49,9 +49,9 @@ In this sense and taking into account the increasing examples of sonification in
 
 ![Graphic User Interface with all the panels visible.\label{fig:fig1}](figures/fig7.png)
 
-![The first framework of the GUI, the data deployed is a galaxy spectrum downloaded from the SDSS database\footnote{https://skyserver.sdss.org/dr12/en/tools/explore/Summary.aspx?ra=179.689293428354&dec=-0.454379056007667}.\label{fig:fig2}](figures/fig6.png)
+![The first framework of the GUI, the data deployed is a galaxy spectrum downloaded from the SDSS database.\label{fig:fig2}](figures/fig6.png)
 
-SonoUno was programmed purely in Python, its modular design with the use of agile methodology, allows programmers to divide the tasks, organize the job and make the cooperative work between the team easier. In this case, the modules are Data Input, Data Output, Data Transformation, Sonification, Graphic User Interface design, and Core. One important remark is the possibility of using each module alone without the need of the graphic user interface (GUI), so some users could import the sound module from bash to produce sonification without installing the whole program.
+SonoUno\footnote{https://skyserver.sdss.org/dr12/en/tools/explore/Summary.aspx?ra=179.689293428354&dec=-0.454379056007667} was programmed purely in Python, its modular design with the use of agile methodology, allows programmers to divide the tasks, organize the job and make the cooperative work between the team easier. In this case, the modules are Data Input, Data Output, Data Transformation, Sonification, Graphic User Interface design, and Core. One important remark is the possibility of using each module alone without the need of the graphic user interface (GUI), so some users could import the sound module from bash to produce sonification without installing the whole program.
 
 Another important point was the actual sound library and implementation on sonoUno, it produces sound from its sinusoidal parameters without any additional feature, each data point on the dataset produces a differentiated tone that faithfully represents the numerical data. In addition, the sonification of each point ensures a good correlation between the visual and sound representation, in consequence, a good match between the two human sensory styles. SonoUno software's principal aim is research through sonification, it is very important to understand and correlate sonification with visualization (the actual practice to research in astronomy) in conjunction with a deeper analysis of human sound perception.
 
