+++ paper/paper.md
@@ -43,13 +43,13 @@ The use of sonification in astronomy has existed for years, some recent examples
 
 Furthermore, @zanella2022, after an international sound workshop held in August 2021, created a repository with existing software until December 2021, which yields a result of 98 projects developed since 1962, many of them discontinued, with a lack of documentation or without evidence of applications in science. Almost 80% of the sonification projects have been carried out between 2011 and 2021. Until 2017, the date on which this sonification software development began, only 50% of the software included in the mentioned repository had started its development.
 
-Nevertheless, in most cases, the sonification mapping of the dataset was defined by its creator and shared as a final product or even with some musicalization, not clearly devoted to research. The ICAD conference intended to bring together multidisciplinary experts working in this field of sonification, since its creation in 1992, in its repository there are a lot of works that @andreopoulouYgoudarzi2021 grouped into six categories (sonification methods, sonification tools/system, review/opinion, exploratory, perception/evaluation studies and other). This systematic review of the ICAD repository highlights a high percentage of papers devoted to sonification methods and tools, in contrast to a low percentage of papers devoted to design methodologies, perception studies, and evaluation methods. In contrast, during The Audible Universe 2, a workshop organized by Lorentz Center\footnote{https://www.lorentzcenter.nl/the-audible-universe-2.html}, the usefulness of evaluation methods and perception studies were pointed out.
+Nevertheless, in most cases, the sonification mapping of the dataset was defined by its creator and shared as a final product or even with some musicalization, not clearly devoted to research. The ICAD conference intended to bring together multidisciplinary experts working in this field of sonification, since its creation in 1992, in its repository there are a lot of works that @andreopoulouYgoudarzi2021 grouped into six categories (sonification methods, sonification tools/system, review/opinion, exploratory, perception/evaluation studies and other). This systematic review of the ICAD repository highlights a high percentage of papers devoted to sonification methods and tools, in contrast to a low percentage of papers devoted to design methodologies, perception studies, and evaluation methods. In contrast, during The Audible Universe 2, a workshop organized by Lorentz Center\footnote{\url{https://www.lorentzcenter.nl/the-audible-universe-2.html}}, the usefulness of evaluation methods and perception studies were pointed out.
 
 In this sense and taking into account the increasing examples of sonification in astrophysics, sonoUno is a sonification software to translate data from two or more column tables into sound.  It is an open-source program with a modular design, that allows users to open different datasets, and explore them through visual and auditory display, the last permitting them to adjust visual and sound settings to enhance their perception (a full view of the Graphic User Interface (GUI) is shown in \autoref{fig:fig1}). This project is user centred from the beginning, to reach that goal at first the ISO 9241-171:2008 standard was used to analyze the accessibility of three previous software [@ise2a2017; @ijskd2022] and construct the first GUI mock-up; second, a theoretical framework centred on visual disability was designed and applied to the software in development [@casadoetal2019], the consequent ISO analysis to sonoUno show very good results. Finally, a focus group session was conducted with people with and without visual disabilities to test the first version of sonoUno, some recommendations and updates arise from that study [@casadoFG2022] (the first view of the GUI after all the updates was shown in \autoref{fig:fig2}).
 
 ![Graphic User Interface with all the panels visible.\label{fig:fig1}](figures/fig7.png)
 
[…]
