+++ paper/paper.md
@@ -65,9 +65,9 @@ In addition to tests inside the research team, sonoUno has been used by other re
 
 \begin{itemize}
   \item the work of Carlos Morales Socorro with variables star, in CEP Las Palmas, Spain, where a training program using sonification was conducted, and a blind student could discover a variable star for the first time in history\footnote{\url{https://astronomiayeducacion.org/taller-2-de-sonificacion-descubriendo-el-universo/?cn-reloaded=1&cn-reloaded=1}};
-  \item the ``Sensing the Dynamic Universe'' project, led by a research group at Harvard [@wandaetal2019], display a web page with information about variables star and use sonoUno to generate the videos of data visualizations and sonification\footenote{\url{https://lweb.cfa.harvard.edu/sdu/rrlyrae.html}}, even in this project a Fork of the sonoUno code was created to make some updates\footenote{\url{https://github.com/joepalmo/sonoUno}};
-  \item the GitHub repository “SonoUno-Raspberry-Pi”\footenote{\url{https://github.com/Physicslibrary/SonoUno-Raspberry-Pi}} of Hartwell Fong, where it is shown the possibility of using sonoUno in a Raspberry Pi, also this user write to us asking for the possibility to include some tactile technology in the software, that could be implemented in a short future;
-  \item ``The Sound of BEARS''\footenote{\url{https://stephenserjeant.github.io/sounds-of-bears/}} is a web page that uses sonoUno web version to display a video with the sonification of ALMA data, reinforcing the need to make astronomy more accessible.
+  \item the ``Sensing the Dynamic Universe'' project, led by a research group at Harvard [@wandaetal2019], display a web page with information about variables star and use sonoUno to generate the videos of data visualizations and sonification\footnote{\url{https://lweb.cfa.harvard.edu/sdu/rrlyrae.html}}, even in this project a Fork of the sonoUno code was created to make some updates\footnote{\url{https://github.com/joepalmo/sonoUno}};
+  \item the GitHub repository “SonoUno-Raspberry-Pi”\footnote{\url{https://github.com/Physicslibrary/SonoUno-Raspberry-Pi}} of Hartwell Fong, where it is shown the possibility of using sonoUno in a Raspberry Pi, also this user write to us asking for the possibility to include some tactile technology in the software, that could be implemented in a short future;
+  \item ``The Sound of BEARS''\footnote{\url{https://stephenserjeant.github.io/sounds-of-bears/}} is a web page that uses sonoUno web version to display a video with the sonification of ALMA data, reinforcing the need to make astronomy more accessible.
 \end{itemize}
 
 # Conclusions
