### 2023-06-28 #5597 @johicasado
@arfon many thanks for your guidance, I already update the text, it is still over 1000 words, it has around 1700, but I change substantially the format and it has now a Statement of need and cases of use, I want to ask if I have to make it shorter or if it could be accepted with this new length.

I don't have any problem to make it shorter if it is needed. I will be waiting for any update. Thanks.
