### 2023-09-08 #5819 @editorialbot
[REVIEW]: SonoUno development: a User Centred Sonification software for data analysis
**Submitting author:** @johicasado (<a href="http://orcid.org/0000-0001-9528-5034">Johanna Casado</a>)
**Repository:** https://github.com/sonoUnoTeam/sonoUno-desktop
**Branch with paper.md** (empty if default branch): master
**Version:** 4.1.1
