### 2023-10-03 #5819 @james-trayford
@johicasado I enjoyed exploring sonoUno and it's functionality, and think it represents a valuable tool in providing accessible sonification of 1D data (in the pitch-mapping tradition) for research applications. I congratulate the developers on their work and hope that the comments  are constructive towards the future development of the code! I think the code works as expected and there are a few relatively minor things to complete the checklist.

I opened a related Issue here: https://github.com/sonoUnoTeam/sonoUno-desktop/issues/11

With some small code suggestions in the PR https://github.com/sonoUnoTeam/sonoUno-desktop/pull/10

My checklist quoted below with some explanations (@ivastar I was unsure if the check means that I have considered that aspect or that aspect has been deemed to already be met by the submission, have assumed the latter)

> ## Review checklist for @james-trayford
> ### Conflict of interest
> * [x]  I confirm that I have read the [JOSS conflict of interest (COI) policy](https://github.com/openjournals/joss/blob/master/COI.md) and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.
>
> ### Code of Conduct
> * [x]  I confirm that I read and will adhere to the [JOSS code of conduct](https://joss.theoj.org/about#code_of_conduct).
>
> ### General checks
> * [x]  **Repository:** Is the source code for this software available at the https://github.com/sonoUnoTeam/sonoUno-desktop?
> * [ ]  **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?

I noticed that the current MIT license includes some extra text, I've included that in the PR https://github.com/sonoUnoTeam/sonoUno-desktop/pull/10

> * [x]  **Contribution and authorship:** Has the submitting author (@johicasado) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
> * [x]  **Substantial scholarly effort:** Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)
> * [x]  **Data sharing:** If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.
> * [x]  **Reproducibility:** If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.
> * [x]  **Human and animal research:** If the paper contains original data research on humans subjects or animals, does it comply with [JOSS's human participants research policy and/or animal research policy](https://joss.readthedocs.io/en/latest/policies.html?highlight=animal#joss-policies)? If the paper contains no such data, please check this item.
>
> ### Functionality
> * [x]  **Installation:** Does installation proceed as outlined in the documentation?
> * [x]  **Functionality:** Have the functional claims of the software been confirmed?
> * [x]  **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
>
> ### Documentation
> * [x]  **A statement of need**: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
> * [x]  **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
> * [x]  **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

it would be good to package the SDSS example data with the code - I had problems with the SDSS server while going through instructions which meant I couldn't proceed. If this is included in the sample data, that could make things more convenient

> * [x]  **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

The User guide document could use some proofreading, but is fully comprehensible. Perhaps some more extension on  the _Octave_ functionality or pointing to external relevant octave documentation  would be helpful.

> * [x]  **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?
> * [x]  **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I think more or less covered, but it might be good to make these three more explicit in e.g. the README

> ### Software paper
> * [x]  **Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
> * [x]  **A statement of need:** Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?
> * [ ]  **State of the field:** Do the authors describe how this software compares to other commonly-used packages?
> * [x]  **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
> * [x]  **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?

The paper is clear and addresses all of these points in my opinion.  One thing I think could be improved is mentioning some other codes that have emerged for sonification  (e.g. `astronify`, `strauss`, `starsound`, and `Desmos` sonification capabilities)  explicitly to better clarify how it fits into the software landscape and emphasise the unique aspects of `sonoUno`. I would think a unique aspect of `sonoUno` is the combination of being a  dedicated GUI-based FOSS and platform independent python code, with the user centered design mentioned already.
