### 2023-10-05 #5819 @ivastar
> My checklist quoted below with some explanations (@ivastar I was unsure if the check means that I have considered that aspect or that aspect has been deemed to already be met by the submission, have assumed the latter)

@james-trayford, yes, this assumption is correct: a check means that the submission meets the requirements
