### 2023-10-11 sonoUnoTeam/sonoUno-desktop#11 @johicasado
Thank you very much @james-trayford for the detailed explanation and all the time dedicated, I appreciate it a lot!

I will put all of these into the agenda, all of the suggestions are very important and some of them also appear in the user testing.

Only some minor comments just in case, about the sort option I think it is great! I don't remember at this moment, which type of data needs to maintain the x value unsorted, but if that problem arises we can just put a function to undo or something like that. About the "I would expect linearly varying in x to be the correct functionality" I will check that. The column names must be included just in case the file doesn't provide one, but maybe it is working badly right now. I will check it.

About the inf's and NaN's values, they are important in some cases, I don't know the better way to proceed with that, my first intuition was to do a survey with the possible option, and evaluate what people need in that cases; about the soft capabilities, I'm thinking using tickmarks to differentiate them from cero, the warning message that you suggest, some section into the configurations; this only some first thoughts.

The GUI requirements and sound production gave me headaches from the beginning. We talked about it in the Audible Universe meeting I think. One option that probably works is to allow the user to deactivate the graph and sound synchronization.

I will work on this topic and post here if I have some news! or I'll open a new issue for a specific exchange.
Thank you very much!
