### 2023-10-13 sonoUnoTeam/sonoUno-desktop#11 @james-trayford
Great, thanks for the explanations! Yeah, some of these questions seem more complicated than others, and I appreciate not all the suggestions may be feasible. In case it wasn't clear, I don't consider addressing all these issues as necessary for the actual JOSS publication, with those specific points addressed in the review thread comment: https://github.com/openjournals/joss-reviews/issues/5819#issuecomment-1745316613
