### 2023-10-26 #5819 @ceb8
@johicasado This is clearly represents substantial effort and is a valuable contribution to the sonification landscape.

I have only minor comments which I intersperse with the relevant checklist items below.

> ## Review checklist for @ceb8
> ### Conflict of interest
>
>     * [x]  I confirm that I have read the [JOSS conflict of interest (COI) policy](https://github.com/openjournals/joss/blob/master/COI.md) and that: I have no COIs with reviewing this work or that any perceived COIs have been waived by JOSS for the purpose of this review.
>
>
> ### Code of Conduct
>
>     * [x]  I confirm that I read and will adhere to the [JOSS code of conduct](https://joss.theoj.org/about#code_of_conduct).
>
>
> ### General checks
>
>     * [x]  **Repository:** Is the source code for this software available at the https://github.com/sonoUnoTeam/sonoUno-desktop?
>
>     * [x]  **License:** Does the repository contain a plain-text LICENSE file with the contents of an [OSI approved](https://opensource.org/licenses/alphabetical) software license?
>
>     * [x]  **Contribution and authorship:** Has the submitting author (@johicasado) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?
>
>     * [x]  **Substantial scholarly effort:** Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)
>
>     * [x]  **Data sharing:** If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.
>
>     * [x]  **Reproducibility:** If the paper contains original results, results are entirely reproducible by reviewers. If the paper contains no original results, please check this item.
>
>     * [x]  **Human and animal research:** If the paper contains original data research on humans subjects or animals, does it comply with [JOSS's human participants research policy and/or animal research policy](https://joss.readthedocs.io/en/latest/policies.html?highlight=animal#joss-policies)? If the paper contains no such data, please check this item.
>
>
> ### Functionality
>
>     * [x]  **Installation:** Does installation proceed as outlined in the documentation?
>
I first installed sonoUno from pip, however that version threw an error, so instead I installed from source which worked perfectly well. Perhaps this means a new pip release is in order.

>     * [x]  **Functionality:** Have the functional claims of the software been confirmed?

A few things I noticed:
- I couldn't really detect a difference between the discrete and continuous sound options, perhaps the continuous notes were a little longer.
-  For large data files even the fastest tempo went through the data quite slowly
- When you make a bunch of panels visible it sometimes causes the layout to do strange things and overlap panels
- Two things that might be because I fed in bad data (see discussion of the requirements on the x-axis below)
    - When you switch the x/y columns weird things can happen,  i.e for `sinsoidal_copia` it plays about 3/4 of the way through, then stops playing sounds while the plot x-axis continually expands to the left
    - The abscissa position goes weird when the data is not sorted on the x-axis column

>
>     * [x]  **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
>
>
> ### Documentation
>
>     * [x]  **A statement of need**: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
>
>     * [x]  **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
>
>     * [x]  **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
>

While everything (installation instructions, user manual, github repo) is nicely linked from the sonoUno website, the website is not mentioned in the GitHub repo. I would be good to include that prominently, right at the top of the README file.  Similarly after the installation instructions in the README it would be good to have a link to the user manual before diving into the bash mode instructions.

Also when the website is mentioned in the user manual itself the URL is not included (or at least in the English version, which was the one I looked at), and should be.

>     * [x]  **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

I don't know what this statement means:

“In the files to import the first column must have continuous values, it is used as a coordinate dependent axis.”

I think it means that the first column needs to be monotonically increasing or decreasing, but it could do with clarification. And if it *does* mean what I think then either the file(s) that do not conform to this format in the `sample_data` directory should be removed (`sinsoidal_copia` is the one I found), or there should be some explanation of the purpose of those example files.

>
>     * [ ]  **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?

The user manual provides much of this functionality, but there are a few holes:
- The average mathematical function is not sufficiently explained, it seems to map the data onto the 0-1 range, and is clearly not a rolling average, but how exactly it is implemented is not clear.
- It's not entirely clear what should happen when bad data is fed into the software

>
>     * [x]  **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
>

In the documentation it indicates that users should communicate any issues to the development team (start of section 2), it would be good to include specific instructions on how.

>
> ### Software paper
>
>     * [x]  **Summary:** Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?
>
>     * [x]  **A statement of need:** Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?
>
>     * [x]  **State of the field:** Do the authors describe how this software compares to other commonly-used packages?
>
>     * [x]  **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
>
>     * [x]  **References:** Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?
