### 2023-11-03 #5819 @johicasado
@ivastar How will we continue the review process? I have already started with the documentation to implement the recommendations. Do I have to indicate here the updates that solve the recommendations? or do I have to open an issue on GitHub?
