### 2023-11-06 #5819 @ivastar
@johicasado you can make all changes to the library, documentation and paper and commit them to the repository. This will allow us to generate a new manuscript here. You don't have to open a new issue. Once the changes are made, please let us know here. The reviewers and I will check the changes.
