### 2023-11-15 99d8196
Update paper.md
Inclusion of other codes devoted to sonification, related to astronomy, answering the recommendations from the paper review.
