+++ paper/paper.md
@@ -45,11 +45,17 @@ Furthermore, @zanella2022, after an international sound workshop held in August 
 
 Nevertheless, in most cases, the sonification mapping of the dataset was defined by its creator and shared as a final product or even with some musicalization, not clearly devoted to research. The ICAD conference intended to bring together multidisciplinary experts working in this field of sonification, since its creation in 1992, in its repository there are a lot of works that @andreopoulouYgoudarzi2021 grouped into six categories (sonification methods, sonification tools/system, review/opinion, exploratory, perception/evaluation studies and other). This systematic review of the ICAD repository highlights a high percentage of papers devoted to sonification methods and tools, in contrast to a low percentage of papers devoted to design methodologies, perception studies, and evaluation methods. In contrast, during The Audible Universe 2, a workshop organized by Lorentz Center\footnote{\url{https://www.lorentzcenter.nl/the-audible-universe-2.html}}, the usefulness of evaluation methods and perception studies were pointed out.
 
-In this sense and taking into account the increasing examples of sonification in astrophysics, sonoUno is a sonification software to translate data from two or more column tables into sound.  It is an open-source program with a modular design, that allows users to open different datasets, and explore them through visual and auditory display, the last permitting them to adjust visual and sound settings to enhance their perception (a full view of the Graphic User Interface (GUI) is shown in \autoref{fig:fig1}). This project is user centred from the beginning, to reach that goal at first the ISO 9241-171:2008 standard was used to analyze the accessibility of three previous software [@ise2a2017; @ijskd2022] and construct the first GUI mock-up; second, a theoretical framework centred on visual disability was designed and applied to the software in development [@casadoetal2019], the consequent ISO analysis to sonoUno show very good results. Finally, a focus group session was conducted with people with and without visual disabilities to test the first version of sonoUno, some recommendations and updates arise from that study [@casadoFG2022] (the first view of the GUI after all the updates was shown in \autoref{fig:fig2}).
+In this sense and taking into account the increasing examples of sonification, only some actual developments related to astronomy, astrophysics, and accessibility will be mentioned here. StarSound started its development at the same time as SonoUno; it offers a complete sound synthesizer customizable by the user. By nature, synthesizers in general present a lot of buttons and require setting a lot of variables, but paying attention to the accessibility of visually impaired people, StarSound allows setting them in a text field too. The IDATA project, beyond their numerous activities centred on accessible astronomical software and materials, launched the Afterglow Access Software, which allows users to open and produce sound from image files\footnote{\url{https://idataproject.org/resources/}}.
 
-SonoUno was programmed purely in Python, its modular design with the use of agile methodology, allows programmers to divide the tasks, organize the job and make the cooperative work between the team easier. In this case, the modules are Data Input, Data Output, Data Transformation, Sonification, Graphic User Interface design, and Core. One important remark is the possibility of using each module alone without the need of the graphic user interface (GUI), so some users could import the sound module from bash to produce sonification without installing the whole program.
+Some other developments center their efforts on sonification and don’t present a graphic user interface; this is the case of Soni-py\footnote{\url{https://github.com/lockepatton/sonipy}}, Astronify\footnote{\url{https://astronify.readthedocs.io/en/latest/}} and STRAUSS\footnote{\url{https://github.com/james-trayford/strauss}}. Soni-py is a Python package devoted to the sonification of scatter plots; it presents examples and demos in its documentation. Astronify allows the sonification of any time series data, which means any data set where the first column was set as time and the second column as pitch; it is an active open-access development with proper documentation and examples with Light Curve.
 
[…]
