### 2023-11-15 4158a6b
Update README.md
Inclution of the webpage links, user manual link and the pip installation was removed because it is not working right now.
