### 2023-12-01 4a85f2d
Updates JOSS review
Following the recommendation of @james-trayford on the JOSS review
(openjournals/joss-reviews#5819), I include the SDSS file used during
the user manual explanations in the sample_data folder.
