### 2023-12-01 e21e213
Bad data erased
Problem detected during the JOSS journal review by @ceb8 on the issue
(openjournals/joss-reviews#5819). The data files are corrupted so how
and was erased to avoid confusion.
