+++ sonoUno/sample_data/sinusoidal_copia.txt
@@ -1,51 +1,0 @@
-abscisa	ordenada
--3.141592653589793116e+00	-1.224646799147353207e-16
--3.013364382014699583e+00	-1.278771616845060521e-01
--2.885136110439606050e+00	-2.536545839095073474e-01
--2.756907838864512517e+00	-3.752670048793740332e-01
--2.628679567289418983e+00	-4.907175520039377403e-01
--2.500451295714325006e+00	-5.981105304912160969e-01
--2.372223024139231473e+00	-6.956825506034864937e-01
--2.243994752564137940e+00	-7.818314824680299147e-01
--2.115766480989044407e+00	-8.551427630053461959e-01
--1.987538209413950874e+00	-9.144126230158124979e-01
--1.859309937838857341e+00	-9.586678530366605777e-01
--1.731081666263763585e+00	-9.871817834144501758e-01
--1.602853394688670052e+00	-9.994862162006878936e-01
--1.474625123113576519e+00	-9.953791129491982304e-01
--1.346396851538482764e+00	-9.749279121818236193e-01
--1.218168579963389231e+00	-9.384684220497604423e-01
--1.089940308388295698e+00	-8.865993063730001067e-01
--9.617120368132021646e-01	-8.201722545969559741e-01
--8.334837652381086315e-01	-7.402779970753157190e-01
--7.052554936630150983e-01	-6.482283953077886407e-01
--5.770272220879215652e-01	-5.455349012105489281e-01
--4.487989505128275880e-01	-4.338837391175581204e-01
--3.205706789377340549e-01	-3.151082180236207120e-01
--1.923424073626405217e-01	-1.911586287013724295e-01
--6.411413578754698861e-02	-6.407021998071310565e-02
-6.411413578754654452e-02 	6.407021998071266156e-02
-1.923424073626400777e-01 	1.911586287013719854e-01
-3.205706789377336108e-01 	3.151082180236203234e-01
-4.487989505128275880e-01 	4.338837391175581204e-01
-5.770272220879211211e-01 	5.455349012105485951e-01
-7.052554936630146543e-01 	6.482283953077883076e-01
-8.334837652381081874e-01 	7.402779970753153860e-01
-9.617120368132017205e-01 	8.201722545969557521e-01
-1.089940308388295698e+00 	8.865993063730001067e-01
-1.218168579963388787e+00 	9.384684220497602203e-01
-1.346396851538482764e+00 	9.749279121818236193e-01
-1.474625123113575853e+00 	9.953791129491981193e-01
-1.602853394688669830e+00 	9.994862162006878936e-01
-1.731081666263762919e+00 	9.871817834144502868e-01
-1.859309937838856897e+00 	9.586678530366606887e-01
-1.987538209413949986e+00 	9.144126230158128310e-01
-2.115766480989043963e+00 	8.551427630053464179e-01
-2.243994752564137940e+00 	7.818314824680299147e-01
-2.372223024139231029e+00 	6.956825506034868267e-01
-2.500451295714325006e+00 	5.981105304912160969e-01
-2.628679567289418095e+00 	4.907175520039385175e-01
-2.756907838864512073e+00 	3.752670048793744773e-01
-2.885136110439605162e+00 	2.536545839095082355e-01
-3.013364382014699139e+00 	1.278771616845064962e-01
-3.141592653589793116e+00 	1.224646799147353207e-16
