+++ sonoUno/sample_data/growing2.txt
@@ -1,71 +1,0 @@
-X	Y
-3.030303030303030454e+01	0.000000000000000000e+00
-3.131313131313131493e+01	1.010101010101010166e+00
-3.232323232323232531e+01	2.020202020202020332e+00
-3.333333333333333570e+01	3.030303030303030276e+00
-3.434343434343434609e+01	4.040404040404040664e+00
-3.535353535353535648e+01	5.050505050505051052e+00
-3.636363636363636687e+01	6.060606060606060552e+00
-3.737373737373737725e+01	7.070707070707070940e+00
-3.838383838383838764e+01	8.080808080808081328e+00
-3.939393939393939803e+01	9.090909090909091717e+00
-4.040404040404040842e+01	1.010101010101010210e+01
-4.141414141414141881e+01	1.111111111111111249e+01
-4.242424242424242919e+01	1.212121212121212110e+01
-4.343434343434343958e+01	1.313131313131313149e+01
-4.444444444444444997e+01	1.414141414141414188e+01
-4.545454545454546036e+01	1.515151515151515227e+01
-4.646464646464647075e+01	1.616161616161616266e+01
-4.747474747474748114e+01	1.717171717171717304e+01
-4.848484848484848442e+01	1.919191919191919382e+01
-4.949494949494949481e+01	2.020202020202020421e+01
-5.050505050505050519e+01	2.121212121212121460e+01
-5.151515151515151558e+01	2.222222222222222499e+01
-5.252525252525252597e+01	2.323232323232323537e+01
-5.353535353535353636e+01	2.424242424242424221e+01
-5.454545454545454675e+01	2.454545454545454675e+01
-5.555555555555555713e+01	2.555555555555555713e+01
-5.656565656565656752e+01	2.656565656565656752e+01
-5.757575757575757791e+01	2.757575757575757791e+01
-5.858585858585858830e+01	2.858585858585858830e+01
-5.959595959595959869e+01	2.959595959595959869e+01
-6.060606060606060908e+01	3.060606060606060908e+01
-6.161616161616161946e+01	3.161616161616161946e+01
-6.262626262626262985e+01	3.262626262626262985e+01
-6.363636363636364024e+01	3.363636363636364024e+01
-6.464646464646465063e+01	3.464646464646465063e+01
-6.565656565656566102e+01	3.565656565656566102e+01
-6.666666666666667140e+01	3.666666666666667140e+01
-6.767676767676768179e+01	3.767676767676768179e+01
-6.868686868686869218e+01	3.868686868686869218e+01
-6.969696969696970257e+01	3.969696969696970257e+01
-7.070707070707071296e+01	4.070707070707071296e+01
-7.171717171717172334e+01	4.171717171717172334e+01
-7.272727272727273373e+01	4.272727272727273373e+01
-7.373737373737374412e+01	4.373737373737374412e+01
-7.474747474747475451e+01	4.474747474747475451e+01
-7.575757575757576490e+01	4.575757575757576490e+01
-7.676767676767677528e+01	4.676767676767677528e+01
-7.777777777777778567e+01	4.777777777777778567e+01
-7.878787878787879606e+01	4.878787878787879606e+01
-7.979797979797980645e+01	4.979797979797980645e+01
-8.080808080808081684e+01	5.080808080808081684e+01
-8.181818181818182723e+01	5.181818181818182723e+01
-8.282828282828283761e+01	5.282828282828283761e+01
-8.383838383838384800e+01	5.383838383838384800e+01
-8.484848484848485839e+01	5.484848484848485839e+01
-8.585858585858586878e+01	5.585858585858586878e+01
-8.686868686868687917e+01	5.686868686868687917e+01
[…]
