### 2023-12-01 5f6c398
User Manual update
Some of the issues highlighted from the @ceb8 and @james-trayford in
their review for JOSS journal (openjournals/joss-reviews#5819) points
out to the User Manual and it was updated.

Changes:
- Include the URL and link to the web page
- Better explanation about average or smooth function (section 4.3.3.1.
Predefined functions)
- Explanation about what type of data the software expect, and example
of wrong data (section 2)
- Contact email and GitHub link included in section 2
- Link to Octave web page included on section 4.3.3.3. Octave
functionality
