### 2023-12-06 #5819 @ivastar
@johicasado I will read the paper throughout and might make some small editorial suggestions by the end of the week. I'll post the suggested edits here. Following that we can proceed with the checklist above.
