+++ paper/paper.bib
@@ -46,6 +46,16 @@
   doi = {10.4018/jitr.2014010105}
 }
 
+@inproceedings{aud,
+  author={Misdariis, N. and Pauletto, S. and Bonne, N. and Harrison, C. and Meredith, K. and Zanella, A.},
+  title={The Audible Universe Workshop: an Interdisciplinary Approach to the Design and Evaluation of Tools for Astronomical Data Sonification},
+  booktitle={Presented at the 28th International Conference on Auditory Displays (ICAD 2023) - Sonification for the Masses, Norrköping, Sweden, 26–30 June 2023},
+  pages={},
+  year={2023},
+  organization={},
+  doi = {}
+}
+
 @article{laurentizetal2021,
   author = {Laurentiz, S. and Bergantini, L. P. and Venancio, S. and Mayer, B. and Carvalho, M. A. and Vargas, D. and Perissinotto, P. M. and Policarpo, C.},
   title = {R Scuti: Creative Star Data Visualization},
