+++ paper/paper.bib
@@ -5,7 +5,7 @@
   year = {2013}
 }
 
-@inproceedings{wandaetal2019, 
+@inproceedings{wanda2019, 
   author = {Díaz-Merced, W. and Walker-Holmes, M. and Hettlage, C. and Green, P. and Casado, J. and García, B.}, 
   title = {A User Centred Inclusive Web Framework for Astronomy}, 
   booktitle={Proceedings of the International Astronomical Union},
