### 2023-12-07 18fbb4e
Update paper.md
Changes made following the recommendations of @ivastar under the review process for JOSS publication
