+++ paper/paper.md
@@ -35,27 +35,27 @@ bibliography: paper.bib
 
 # Summary
 
-In general, people think about information and nature based on a visual approach, but if we go back to our childhood and try to understand the world through our different senses, a new way of understanding the information arises. Even when it’s challenging to think of a method to understand astrophysics through other sensorial inputs, sonification is a technique that uses non-speech audio to translate data into sound. The sonification field has grown continuously over the recent years, especially since 1992 with the International Community for Auditory Display-ICAD conferences. In this contribution, a sonification software, sonoUno, and its application in astronomy are presented, without neglecting the analysis possibilities for data from various disciplines. The differences between the present development from others are: it is user centred from the beginning, with several contacts with end users along the design and a focus group test; it seeks a way to bring autonomy to its functional diverse users; and it presents a simple first framework, with the mathematical operations and settings in the menu. A description of the software and some cases of use are presented.
+In general, people think about information and nature based on a visual approach, but if we go back to our childhood and try to understand the world through our different senses, a new way of understanding the information arises. Sonification is a technique that uses non-speach audio to translate data into sound and it is a commonly-used approach to exploring astrophysical data through sensory inputs. This field has grown continuously over the recent years, especially since 1992 with the start of the International Community for Auditory Display (ICAD) conferences. In this contribution, a sonification software, sonoUno, and its application in astronomy are presented, without neglecting the analysis possibilities for data from various disciplines. The main distinguishing features of the development presented here are: it is user centred from the beginning, with several contacts with end users along the design and a focus group test; it seeks a way to bring autonomy to its functionally-diverse users; and it presents a simple first framework, with the mathematical operations and settings in the menu. A description of the software and some cases of use are presented.
 
 # Introduction
 
 The use of sonification in astronomy has existed for years, some recent examples are a sonification of the zCOSMOS astronomical dataset, where the authors describe the dataset and the sonification strategy used [@bardellietal2021]; the LightSound, an electronic device that allows the conversion of light into sound and it is used to observe eclipses [@lightsound2020]; See-Through-Sound, a project dedicated to converting images into sound to allow visually impaired people to detect objects around them [@henriquesetal2014]; R-Scuti, an audio-visual installation that proposes the conversion of astronomical data from the AAVSO (American Association of Variable Star Observers) database into an exhibition environment, using recordings of variable star observations [@laurentizetal2021]; and an accessible science lab, research that highlights the need for multimodal approaches in teaching and learning environments, presenting principles for inclusive material design based on user-centred design and universal design for teaching environments [@reynagaetal2020].
 
-Furthermore, @zanella2022, after an international sound workshop held in August 2021, created a repository with existing software until December 2021, which yields a result of 98 projects developed since 1962, many of them discontinued, with a lack of documentation or without evidence of applications in science. Almost 80% of the sonification projects have been carried out between 2011 and 2021. Until 2017, the date on which this sonification software development began, only 50% of the software included in the mentioned repository had started its development.
+After an international sonification workshop held in August of 2021, @zanella2022 created a repository of existing software as of December 2021 which contained 98 projects developed since 1962. Unfortunately many of them were no longer actively developed, lacked documentation and had no evidence of science applications. Almost 80% of the sonification projects have been carried out between 2011 and 2021. As of 2017, the date on which this sonification software development began, only 50% of the software included in the mentioned repository had started its development.
 
[…]
