+++ paper/paper.bib
@@ -53,7 +53,7 @@
   pages={},
   year={2023},
   organization={},
-  doi = {}
+  doi = {10.21785/icad2023.2124}
 }
 
 @article{laurentizetal2021,
