### 2023-12-07 #5819 @johicasado
@ivastar I added the missing reference to bib file
